// database.types.ts
export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json | undefined }
  | Json[]

export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "13.0.5"
  }
  public: {
    Tables: {
      band_user: {
        Row: {
          band_id: number | null
          id: number
          user_id: number | null
        }
        Insert: {
          band_id?: number | null
          id?: number
          user_id?: number | null
        }
        Update: {
          band_id?: number | null
          id?: number
          user_id?: number | null
        }
        Relationships: [
          {
            foreignKeyName: "fk_band_user_bands"
            columns: ["band_id"]
            isOneToOne: false
            referencedRelation: "bands"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "fk_band_user_users"
            columns: ["user_id"]
            isOneToOne: false
            referencedRelation: "users"
            referencedColumns: ["id"]
          },
        ]
      }
      bands: {
        Row: {
          archived: boolean | null
          creator_user_id: number | null
          end_date: string | null
          end_time: string | null
          id: number
          name: string | null
          start_date: string | null
          start_time: string | null
          token: string | null
        }
        Insert: {
          archived?: boolean | null
          creator_user_id?: number | null
          end_date?: string | null
          end_time?: string | null
          id?: number
          name?: string | null
          start_date?: string | null
          start_time?: string | null
          token?: string | null
        }
        Update: {
          archived?: boolean | null
          creator_user_id?: number | null
          end_date?: string | null
          end_time?: string | null
          id?: number
          name?: string | null
          start_date?: string | null
          start_time?: string | null
          token?: string | null
        }
        Relationships: []
      }
      fixed_schedules: {
        Row: {
          schedule: Json
          user_id: number
        }
        Insert: {
          schedule: Json
          user_id?: number
        }
        Update: {
          schedule?: Json
          user_id?: number
        }
        Relationships: []
      }
      schedules: {
        Row: {
          band_id: number | null
          comment: string | null
          id: number
          schedule: Json | null
          user_id: number | null
        }
        Insert: {
          band_id?: number | null
          comment?: string | null
          id?: number
          schedule?: Json | null
          user_id?: number | null
        }
        Update: {
          band_id?: number | null
          comment?: string | null
          id?: number
          schedule?: Json | null
          user_id?: number | null
        }
        Relationships: []
      }
      users: {
        Row: {
          email: string | null
          id: number
          name: string | null
        }
        Insert: {
          email?: string | null
          id?: number
          name?: string | null
        }
        Update: {
          email?: string | null
          id?: number
          name?: string | null
        }
        Relationships: []
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}

type DatabaseWithoutInternals = Omit<Database, "__InternalSupabase">

type DefaultSchema = DatabaseWithoutInternals[Extract<keyof Database, "public">]
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends
    | keyof DefaultSchema["Enums"]
    | { schema: keyof DatabaseWithoutInternals },
  EnumName extends DefaultSchemaEnumNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = DefaultSchemaEnumNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema["Enums"]
    ? DefaultSchema["Enums"][DefaultSchemaEnumNameOrOptions]
    : never

export type CompositeTypes<
  PublicCompositeTypeNameOrOptions extends
    | keyof DefaultSchema["CompositeTypes"]
    | { schema: keyof DatabaseWithoutInternals },
  CompositeTypeName extends PublicCompositeTypeNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"]
    : never = never,
> = PublicCompositeTypeNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"][CompositeTypeName]
  : PublicCompositeTypeNameOrOptions extends keyof DefaultSchema["CompositeTypes"]
    ? DefaultSchema["CompositeTypes"][PublicCompositeTypeNameOrOptions]
    : never